It no longer makes sense to have an abstract JoinCommandRunner
The JoinCommand type is responsible for managing different ways of running commands, so any differences in how commands are run should be handled by it, rather than the runner.
