--- src/main/java/com/alexiwolf/joincommands/SimpleJoinCommandRunner.java
@@ -1,25 +1,0 @@
-package com.alexiwolf.joincommands;
-
-import org.bukkit.entity.Player;
-
-import java.util.List;
-
-public class SimpleJoinCommandRunner extends JoinCommandRunner {
-    private final List<String> newPlayerCommands;
-    private final List<String> returningPlayerCommands;
-
-    public SimpleJoinCommandRunner(List<String> newPlayerCommands, List<String> returningPlayerCommands) {
-        this.newPlayerCommands = newPlayerCommands;
-        this.returningPlayerCommands = returningPlayerCommands;
-    }
-
-    @Override
-    public void runCommandsForNewPlayer(Player player) {
-        newPlayerCommands.forEach(player::performCommand);
-    }
-
-    @Override
-    public void runCommandsForReturningPlayer(Player player) {
-        returningPlayerCommands.forEach(player::performCommand);
-    }
-}
